Finish adding last commandline arg and update README.

+++ Assignment6_SceneGraphOrParticleSystem/ParticleSystem/FountainEmitter.cpp
@@ -10,7 +10,6 @@ FountainEmitter::FountainEmitter(Camera* camera, const QVector3D& position, cons
 }
 
 void FountainEmitter::update(unsigned int msSinceLastFrame) {
-//  qDebug() << "msSinceLastFrame " << msSinceLastFrame;
 	timeLivedSoFar += msSinceLastFrame;
 	particleModel_->update(msSinceLastFrame);
 	// remove dead particles
@@ -22,7 +21,7 @@ void FountainEmitter::update(unsigned int msSinceLastFrame) {
     }
 	
 	// add new particle
-	int particlesRate = 1000.0 / particlesPerSecond_;
+	float particlesRate = 1000.0 / particlesPerSecond_;
 	int particlesToAdd = (timeLivedSoFar - lastEmittedAt) / particlesRate;
 	for(int i = 0; i < particlesToAdd; i++) {
 		if (i == 0) lastEmittedAt = timeLivedSoFar;
@@ -30,11 +29,6 @@ void FountainEmitter::update(unsigned int msSinceLastFrame) {
 		this->particles_.append(newP);
 	}
 
-//	qDebug() << "removedParticles " << removedParticles;
-//	qDebug() << "particlesToAdd " << particlesToAdd;
-//	qDebug() << "particles_.size() " << particles_.size();
-//
-	// update all particles
     for (auto p: particles_) {
     	p->updateAndDraw(msSinceLastFrame);
 	}
